Move collection filename schema to common package
