+++ galaxy/common/schema.py
@@ -14,8 +14,13 @@
 #
 # You should have received a copy of the Apache License
 # along with Galaxy.  If not, see <http://www.apache.org/licenses/>.
+import re
 
+import attr
 from marshmallow import fields
+import semantic_version as semver
+
+from galaxy import constants
 
 
 class Enum(fields.Field):
@@ -30,3 +35,36 @@ class Enum(fields.Field):
         if self.allow_none and value is None:
             return None
         return self.enum(value).value
+
+
+_FILENAME_RE = re.compile(
+    r'^(?P<namespace>\w+)-(?P<name>\w+)-'
+    r'(?P<version>[0-9a-zA-Z.+-]+)\.tar\.gz$'
+)
+
+
+@attr.s(slots=True)
+class CollectionFilename(object):
+
+    namespace = attr.ib()
+    name = attr.ib()
+    version = attr.ib(converter=semver.Version)
+
+    @classmethod
+    def parse(cls, filename):
+        match = _FILENAME_RE.match(filename)
+        if not match:
+            raise ValueError(
+                'Invalid filename. Expected: '
+                '{namespace}-{name}-{version}.tar.gz'
+            )
+
+        return cls(**match.groupdict())
+
+    @namespace.validator
+    @name.validator
+    def _validator(self, attribute, value):
+        if not constants.NAME_REGEXP.match(value):
+            raise ValueError(
+                'Invalid {0}: {1!r}'.format(attribute.name, value)
+            )
